Rename antivirus client module
